requirements.txt
paho-mqtt
requests
python-dotenv
colorama
flask